MAINT: move MatVecOperator to tensor_ops module
